+++ docs/conf.py
@@ -18,9 +18,10 @@ import os
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
-sys.path.insert(0, os.path.abspath('../wscheck'))
+for path in ('utils', '..'):
+    sys.path.insert(0, os.path.abspath(path))
 
-from version import Version  # noqa: E402
+from wscheck.version import Version  # noqa: E402
 version_info = Version()
 
 # Read the Docs - https://readthedocs.org/
@@ -37,9 +38,10 @@ on_rtd = os.environ.get('READTHEDOCS', None) == 'True'
 # ones.
 extensions = [
     # 'sphinx.ext.autodoc',
-    # 'sphinx.ext.doctest',
+    'sphinx.ext.doctest',
     # 'sphinx.ext.todo',
     # 'sphinx.ext.viewcode',
+    'rules',
 ]
 
 # Add any paths that contain templates here, relative to this directory.