+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ## [Unreleased][unreleased]
+### Added
+- Added PyPy support
+
 ### Fixed
 - Fixed the HTML generation error of readme on GitHub and PyPi