+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ## [Unreleased][unreleased]
+
+
+## [1.1.0] - 2017-11-02
 ### Added
 - Added rule for checking empty lines at the beginning of the files ([WSC007])
 - Added PyPy support
@@ -77,7 +80,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   [WSW005][WSC005], [WSW006][WSC006] rules)
 
 
-[unreleased]: https://github.com/andras-tim/wscheck/compare/v1.0.1...HEAD
+[unreleased]: https://github.com/andras-tim/wscheck/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/andras-tim/wscheck/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/andras-tim/wscheck/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/andras-tim/wscheck/compare/v0.1.6...v1.0.0
 [0.1.6]: https://github.com/andras-tim/wscheck/compare/v0.1.5...v0.1.6